security-module/src/main/java/com/ca/mfaas/security/config/BeansConfiguration.java
/*
 * This program and the accompanying materials are made available under the terms of the
 * Eclipse Public License v2.0 which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-v20.html
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Copyright Contributors to the Zowe Project.
 */
package com.ca.mfaas.security.config;

import com.ca.mfaas.error.ErrorService;
import com.ca.mfaas.error.impl.ErrorServiceImpl;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

@Configuration
public class BeansConfiguration {
    @Bean(name = "securityErrorService")
    public ErrorService errorService() {
        return new ErrorServiceImpl("/security-messages.yml");
    }

    @Bean(name = "securityObjectMapper")
    public ObjectMapper objectMapper() {
        return new ObjectMapper();
    }
}
